v1.4
Forecast for Today radi, postoje jos neki mali bagovi oko lokacije

+++ app/src/main/java/com/prvi/weatherapp/Common/Common.java
@@ -10,14 +10,14 @@ public class Common {
     public static final String APP_ID = "5d36b5d8c14c7600246ef24ae19620bd";
     public static Location current_location = null;
 
-    public static String convertUnixToDate(int dt) {
+    public static String convertUnixToDate(long dt) {
         Date date = new Date(dt*1000L);
         SimpleDateFormat sdf = new SimpleDateFormat("HH:mm EEE MM yyyy");
         String formatted = sdf.format(date);
         return formatted;
     }
 
-    public static String convertUnixToHour(int dt) {
+    public static String convertUnixToHour(long dt) {
         Date date = new Date(dt*1000L);
         SimpleDateFormat sdf = new SimpleDateFormat("HH:mm");
         String formatted = sdf.format(date);
